AnaliseEstoqueXCusto/requirements.txt
pyodbc==5.0.1
pandas==2.1.4
python-dotenv==1.0.0
Flask==3.0.0
pandas
pyodbc
python-dotenv
